refactor:重构点赞逻辑

+++ repository/articles_relation_like_repository.go
@@ -12,6 +12,7 @@ type ArticlesRelationLikeRepository interface {
 	Create(ctx context.Context, articleRelationLike *model.ArticlesRelationLike) error
 	Delete(ctx context.Context, articleRelationLike *model.ArticlesRelationLike) error
 	Get(ctx context.Context, articleRelationLike *model.ArticlesRelationLike) bool
+	List(ctx context.Context, articleID int64) (int64, error)
 }
 
 type articlesRelationLikeRepository struct {
@@ -31,11 +32,17 @@ func (r *articlesRelationLikeRepository) Create(ctx context.Context, articleRela
 }
 
 func (r *articlesRelationLikeRepository) Delete(ctx context.Context, articleRelationLike *model.ArticlesRelationLike) error {
-	return r.DB.Delete(articleRelationLike).Error
+	return r.DB.Delete(articleRelationLike, "article_id = ? AND user_id = ?", articleRelationLike.ArticleID, articleRelationLike.UserID).Error
 }
 
 func (r *articlesRelationLikeRepository) Get(ctx context.Context, articleRelationLike *model.ArticlesRelationLike) bool {
 	var count int64
-	r.DB.Model(articleRelationLike).Where(articleRelationLike).Count(&count)
+	r.DB.Model(articleRelationLike).Where("article_id = ? AND user_id = ?", articleRelationLike.ArticleID, articleRelationLike.UserID).Count(&count)
 	return count > 0
+}
+
+func (r *articlesRelationLikeRepository) List(ctx context.Context, articleID int64) (int64, error) {
+	var articlesRelationLikesCount int64
+	r.DB.Model(&model.ArticlesRelationLike{}).Where("article_id = ?", articleID).Count(&articlesRelationLikesCount)
+	return articlesRelationLikesCount, nil
 }
